InitClose.h eta InitClose.h .c sortu dira, bertan hasieratu eta amaitzeko funtzioak gehituko dira.
